fix subscription callback invoke and pubsub tests
The subscription look callbacks were getting invoked with the `SubscriptionMessage` struct. But the docs do not mention that, and seem to suggest that the callback would receive the actual message, which seems natural. The callbacks in tests seem to be expecting the actual message too.

Updated the callback invocation to pass the actual message instead of `SubscriptionMessage` struct.